+++ src/position.rs
@@ -75,8 +75,8 @@ impl Position {
 
     pub(crate) fn to_vscode_position(self) -> ki_protocol_types::Position {
         ki_protocol_types::Position {
-            character: self.column,
-            line: self.line,
+            character: self.column as u32,
+            line: self.line as u32,
         }
     }
 }
