+++ src/vscode/handlers/ping.rs
@@ -5,7 +5,7 @@ use crate::vscode::app::VSCodeApp;
 use ki_protocol_types::OutputMessage;
 
 /// Handle ping requests from VSCode
-pub fn handle_ping_request(app: &mut VSCodeApp, id: u64, value: String) -> Result<()> {
+pub fn handle_ping_request(app: &mut VSCodeApp, id: u32, value: String) -> Result<()> {
     info!("Received ping: {}", value);
     let response_message = format!("pong: {}", value);
     app.send_response(id, OutputMessage::Ping(response_message))
